DailyReward 플러그인 구조 수정
