+++ src/main/java/net/teamuni/dailyreward/gui/RewardManager.java
@@ -1,14 +1,13 @@
-package net.teamuni.dailyreward;
+package net.teamuni.dailyreward.gui;
 
+import net.teamuni.dailyreward.Dailyreward;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.Material;
 import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.configuration.file.YamlConfiguration;
-import org.bukkit.entity.Player;
 import org.bukkit.event.Listener;
-import org.bukkit.inventory.Inventory;
 import org.bukkit.inventory.ItemFlag;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.ItemMeta;
@@ -18,9 +17,13 @@ import java.io.File;
 import java.util.*;
 
 public class RewardManager implements Listener {
-    private final Dailyreward main = Dailyreward.getPlugin(Dailyreward.class);
-    private File file = null;
+    private final Dailyreward main;
+    private File file;
     private FileConfiguration rewardsFile = null;
+
+    public RewardManager(Dailyreward instance) {
+        this.main = instance;
+    }
 
     public void createRewardsYml() {
         this.file = new File(main.getDataFolder(), "rewards.yml");
@@ -29,41 +32,36 @@ public class RewardManager implements Listener {
         }
     }
 
+    public void reloadRewardsYml() {
+        this.rewardsFile = YamlConfiguration.loadConfiguration(file);
+    }
+
+    public ConfigurationSection loadConfigurationSection() {
+        return rewardsFile.getConfigurationSection("Rewards");
+    }
+
     public FileConfiguration getRewardsFile() {
         return this.rewardsFile;
     }
 
-    public void reload() {
-        this.rewardsFile = YamlConfiguration.loadConfiguration(file);
-    }
-
-    private FileConfiguration getPlayerFileConfiguration(UUID uuid) {
-        File file = new File("plugins/Dailyreward/Players", uuid + ".yml");
-        return YamlConfiguration.loadConfiguration(file);
+    public ConfigurationSection getSection(String key) {
+        return loadConfigurationSection().getConfigurationSection(key);
     }
 
     public int getKeyDay(String key) {
         return Integer.parseInt(key.replaceAll("\\D", ""));
     }
 
-    public int getPlayerCumulativeDate(UUID uuid) {
-        return getPlayerFileConfiguration(uuid).getInt("CumulativeDate");
-    }
-
-    public List<String> getPlayerReceivedRewardsList(UUID uuid) {
-        return getPlayerFileConfiguration(uuid).getStringList("ReceivedRewards");
-    }
-
     private List<String> rewardLoreSet(ConfigurationSection section, String key, UUID uuid) {
         List<String> rewardLoreList = new ArrayList<>();
         for (String lore : section.getStringList("lore")) {
             if (lore.contains("%rewards_receipt_status%")) {
-                if (getKeyDay(key) > getPlayerCumulativeDate(uuid)) {
+                if (getKeyDay(key) > main.getPlayerDataManager().getPlayerCumulativeDate(uuid)) {
                     String placeholderLore = lore.replace("%rewards_receipt_status%", "아직 해당 일차보상을 획득할 수 없습니다.");
                     rewardLoreList.add(ChatColor.translateAlternateColorCodes('&', placeholderLore));
                 } else {
                     String placeholderLore;
-                    if (getPlayerReceivedRewardsList(uuid).contains(key)) {
+                    if (main.getPlayerDataManager().getPlayerReceivedRewardsList(uuid).contains(key)) {
                         placeholderLore = lore.replace("%rewards_receipt_status%", "이미 해당 일차보상을 수령했습니다.");
                     } else {
                         placeholderLore = lore.replace("%rewards_receipt_status%", "해당 일차보상을 수령할 수 있습니다.");
@@ -100,7 +98,7 @@ public class RewardManager implements Listener {
 
     @NotNull
     public Map<Integer, ItemStack> getRewards(UUID uuid) {
-        getPlayerFileConfiguration(uuid);
+        main.getPlayerDataManager().getPlayerFileConfiguration(uuid);
         ConfigurationSection section = this.rewardsFile.getConfigurationSection("Rewards");
         Map<Integer, ItemStack> rewards = new HashMap<>();
         Set<String> rewardsKeys = section.getKeys(false);
@@ -115,18 +113,4 @@ public class RewardManager implements Listener {
         }
         return rewards;
     }
-
-    private void loadItems(Inventory gui, Player player) {
-        Map<Integer, ItemStack> dailyItem = new HashMap<>(getRewards(player.getUniqueId()));
-        for (Map.Entry<Integer, ItemStack> dailyItems : dailyItem.entrySet()) {
-            gui.setItem(dailyItems.getKey(), dailyItems.getValue());
-        }
-    }
-
-    public void openGui(Player player) {
-        Inventory dailyRewardGui = Bukkit.createInventory(null, 54, ChatColor.GREEN + "출석체크 GUI");
-        loadItems(dailyRewardGui, player);
-        player.openInventory(dailyRewardGui);
-    }
 }
-
